refactor(api): extract schedules router (pure move)
Move the 11 schedule routes (list/create, active/page, validate,
default-page, enabled, get/update/delete) plus their two domain-local
helpers (_enrich_schedule_with_sun_times, _with_compat_warnings) from
api_server.py to src/schedules/routes.py, included at the block's old
position so registration order (specific before parameterized) is
unchanged. Handler bodies are verbatim; only imports were adjusted.

api_server-owned names (service getters, is_collection_id, the
_temporary_override_payload/_resolve_* trio shared with the settings
routes, check_ref_board_compatibility) resolve through src.api_server
at call time — the suite patches them there. get_schedule_service is
now unused inside api_server but kept imported (noqa: F401) as that
patch seam.

Response-shape goldens: zero changes. Route-inventory golden: zero
changes. /settings/silence-schedule and /silence-status are settings
domain and stay put.

Part of #1849.
